requirements.txt
numpy==1.26.4
pillow==12.1.1
